added middleware to get token and restrict delete blog to only authorised users

+++ part4/bloglist/controllers/blogs.js
@@ -3,14 +3,6 @@ const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
-
 blogsRouter.get('/', async (request, response, next) => {
   try {
     const blogs = await Blog.find({}).populate('user', {
@@ -25,11 +17,9 @@ blogsRouter.get('/', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body;
-  const token = getTokenFrom(request);
-
   try {
-    const decodedToken = jwt.verify(token, process.env.SECRET);
-    if (!token || !decodedToken.id) {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!request.token || !decodedToken.id) {
       return response
         .status(401)
         .json({ error: 'token missing token or invalid' });
@@ -58,9 +48,25 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = request.token;
+
   try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!token || !decodedToken) {
+      return response
+        .status(401)
+        .json({ error: 'token missing token or invalid' });
+    }
+
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(request.params.id);
+    const userId = user._id;
+    if (userId.toString() === blog.user.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      response.status(401).end();
+    }
   } catch (exception) {
     next(exception);
   }
